+++ TravelDiscovery/Discover/PopularDestinationsView.swift
@@ -47,25 +47,13 @@ struct PopularDestinationsView: View {
                                 .padding(.bottom, 8)
                                 .foregroundColor(.gray)
                         }
-                        .modifier(TileModifier())
+                        .asTile()
                         .padding(.bottom)
                     }
                 }
                 .padding(.horizontal)
             }
         }
-    }
-}
-
-struct TileModifier: ViewModifier {
-    func body(content: Content) -> some View {
-        content
-            .background(Color.white)
-            .cornerRadius(5)
-            .shadow(color: .init(.sRGB, white: 0.8, opacity: 1),
-                    radius: 4,
-                    x: 0,
-                    y: 2)
     }
 }
